+++ Packages/com.nickmaltbie.openkcc.netcode/OpenKCC.netcode/Environment/NetworkMovingPlatform.cs
@@ -18,20 +18,38 @@
 
 using System.Collections.Generic;
 using nickmaltbie.TestUtilsUnity;
+using Unity.Netcode;
 using UnityEngine;
 
-namespace nickmaltbie.OpenKCC.Environment
+namespace nickmaltbie.OpenKCC.netcode.Environment
 {
-    /// <summary>
-    /// Script to translate a rigidbody object between two positions.
-    /// </summary>
     [RequireComponent(typeof(Rigidbody))]
-    public class MovingPlatform : MonoBehaviour
+    public class NetworkMovingPlatform : NetworkBehaviour
     {
+        /// <summary>
+        /// Moving platform with network configuration.
+        /// </summary>
+        private NetworkVariable<int> _currentTarget = new NetworkVariable<int>(
+            value: 0,
+            readPerm: NetworkVariableReadPermission.Everyone,
+            writePerm: NetworkVariableWritePermission.Server);
+
+        /// <summary>
+        /// Is continous variable with network configuration.
+        /// </summary>
+        private NetworkVariable<bool> _isContinuous = new NetworkVariable<bool>(
+            value: true,
+            readPerm: NetworkVariableReadPermission.Everyone,
+            writePerm: NetworkVariableWritePermission.Server);
+
         /// <summary>
         /// Current target the platform is heading for.
         /// </summary>
-        public int currentTargetIndex { get; private set; } = 0;
+        public int CurrentTargetIdx
+        {
+            get => _currentTarget.Value;
+            private set => _currentTarget.Value = value;
+        }
 
         /// <summary>
         /// Unity service for managing time.
@@ -43,7 +61,11 @@ namespace nickmaltbie.OpenKCC.Environment
         /// or discrete steps. uses the MovePosition and MoveRotation api
         /// for continuous movement otherwise.
         /// </summary>
-        internal bool isContinuous = true;
+        internal bool IsContinuous
+        {
+            get => _isContinuous.Value;
+            set => _isContinuous.Value = value;
+        }
 
         /// <summary>
         /// Velocity at which this platform should move.
@@ -57,12 +79,12 @@ namespace nickmaltbie.OpenKCC.Environment
         /// </summary>
         [SerializeField]
         [Tooltip("List of targets to move between.")]
-        internal List<Transform> targetsList;
+        internal List<Transform> targetsList = new List<Transform>();
 
         /// <summary>
         /// Gets the current target we're moving towards.
         /// </summary>
-        public Transform CurrentTarget => ValidTarget ? targetsList[currentTargetIndex] : null;
+        public Transform CurrentTarget => ValidTarget ? targetsList[CurrentTargetIdx] : null;
 
         /// <summary>
         /// Does the moving platform have a valid target.
@@ -71,7 +93,7 @@ namespace nickmaltbie.OpenKCC.Environment
 
         public void FixedUpdate()
         {
-            if (!ValidTarget)
+            if (!ValidTarget || !IsServer)
             {
                 return;
             }
@@ -85,10 +107,10 @@ namespace nickmaltbie.OpenKCC.Environment
             if (direction == Vector3.zero || distanceToTarget < displacement.magnitude)
             {
                 displacement = CurrentTarget.position - rb.position;
-                currentTargetIndex = (currentTargetIndex + 1) % targetsList.Count;
+                CurrentTargetIdx = (CurrentTargetIdx + 1) % targetsList.Count;
             }
 
-            if (isContinuous)
+            if (IsContinuous)
             {
                 rb.MovePosition(rb.position + displacement);
             }